final variable examples

+++ 02/demos/01-lambda-expressions/src/main/java/com/pluralsight/streamslambdas/Product.java
@@ -1,7 +1,6 @@
 
 package com.pluralsight.streamslambdas;
 
-import java.math.BigDecimal;
 import java.util.Locale;
 import java.util.Objects;
 
@@ -9,9 +8,9 @@ public class Product {
 
     private final Category category;
     private final String name;
-    private final BigDecimal price;
+    private final Double price;
 
-    public Product(Category category, String name, BigDecimal price) {
+    public Product(Category category, String name, Double price) {
         this.category = category;
         this.name = name;
         this.price = price;
@@ -25,7 +24,7 @@ public class Product {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
